refactor!: Refactored and cleaned the whole PR
