+++ docs/experimental/viz_laplacian.py
@@ -2,7 +2,6 @@ import numpy as np
 
 from fury.actor import cube, sphere
 from fury.actors.effect_manager import EffectManager
-from fury.shaders import shader_apply_effects, shader_custom_uniforms
 from fury.window import (Scene, ShowManager, record)
 
 def normalize(array : np.array, min : float = 0.0, max : float = 1.0, axis : int = 0):
@@ -59,7 +58,7 @@ manager.scene.add(gauss_cube)
 
 # effects.remove_effect(gauss_cube)
 
-interactive = False
+interactive = True
 
 if interactive:
     manager.start()
